requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pandas
openpyxl
fpdf
